Feat: increase jwt token expiration time
Increase jwt token expiration time from 7 days to 30 days

+++ src/main/java/nl/app/loopsnelheid/security/config/JwtUtils.java
@@ -19,7 +19,7 @@ public class JwtUtils
     private String jwtSecret;
 
     @Value("${security.jwt.expiration-in-ms}")
-    private int jwtExpiration;
+    private Long jwtExpiration;
 
     public String generateJwtToken(Authentication authentication)
     {
